+++ src/Pin/Reminder/Generic.php
@@ -40,7 +40,7 @@ class Generic extends Base
      */
     public function setLocationName($locationName)
     {
-        $this->locationName = $locationName;
+        $this->locationName = str_limit($locationName, 256);
 
         return $this;
     }